Refactor util code into utils module
