+++ processes/utils.py
@@ -1,0 +1,32 @@
+import re
+import xarray as xr
+
+
+def create_product_measurement(platform, all_measurements):
+
+    if platform in ["SENTINEL_2"]:
+        product = "s2_esa_sr_granule"
+        measurements = all_measurements + ["coastal_aerosol", "scene_classification"]
+        # Change with S2 WOFS ready
+        water_product = "SENTINEL_2_PRODUCT DEFS"
+    else:
+        product_match = re.search("LANDSAT_(\d)", platform)
+        if product_match:
+            product = f"ls{product_match.group(1)}_usgs_sr_scene"
+            measurements = all_measurements + ["pixel_qa"]
+            water_product = f"ls{product_match.group(1)}_water_classification"
+        else:
+            raise Exception(f"invalid platform_name {platform}")
+
+    return product, measurements, water_product
+
+
+def is_dataset_empty(ds: xr.Dataset) -> bool:
+    checks_for_empty = [
+        lambda x: len(x.dims) == 0,  # Dataset has no dimensions
+        lambda x: len(x.data_vars) == 0,  # Dataset no variables
+    ]
+    for f in checks_for_empty:
+        if f(ds):
+            return True
+    return False
